requirements.txt
msgpack_python==0.5.6
PyYAML==6.0
xmltodict==0.13.0
